+++ atxm/tx.py
@@ -185,32 +185,3 @@ def _deserialize_tx_receipt(receipt: Dict) -> TxReceipt:
             "status": receipt["status"],
         }
     )
-
-
-def _make_tx_params(data: TxData) -> TxParams:
-    """
-    TxData -> TxParams: Creates a transaction parameters
-    object from a transaction data object for broadcast.
-
-    This operation is performed in order to "turnaround" the transaction
-    data object as queried from the RPC provider (eth_getTransaction) into a transaction
-    parameters object for strategics and re-broadcast (LocalAccount.sign_transaction).
-    """
-    params = TxParams(
-        {
-            "nonce": data["nonce"],
-            "chainId": data["chainId"],
-            "gas": data["gas"],
-            "to": data["to"],
-            "value": data["value"],
-            "data": data.get("data", b""),
-        }
-    )
-    if "gasPrice" in data:
-        params["type"] = "0x01"
-        params["gasPrice"] = data["gasPrice"]
-    elif "maxFeePerGas" in data:
-        params["type"] = "0x02"
-        params["maxFeePerGas"] = data["maxFeePerGas"]
-        params["maxPriorityFeePerGas"] = data["maxPriorityFeePerGas"]
-    return params
